Update action using with MainScreen & query debounce

+++ app/src/main/java/com/example/pixbayphoto/presentation/screen/main/MainEvent.kt
@@ -2,5 +2,5 @@ package com.example.pixbayphoto.presentation.screen.main
 
 sealed interface MainEvent {
     data class ShowSnackbar(val message: String) : MainEvent
-    data class NavigateToDetail(val id: Int) : MainEvent
+    data class NavigateToDetail(val id: Long) : MainEvent
 }
